Cleaned up the html and tried to add the logo.
Cleaned up the html by putting everything that was created earlier into a new vue component. This is to not make it more cluttered than it already is. Then I also tried to add the logo but the "flex-grow: 1" won't really work even though the parent is set to relative. Very strange. Problem for another day!