Fixed the logo!
The problem with the logo was that by default using flex, the items can't shrink below their original size, which means that the image (600 by 600 px) wouldn't shrink below it's size creating this massive icon. The solution was to just add the basic css: "min-width: 0; min-height: 0;". This resolved in the attributes being overwritten and thus working! OMG! I really banged my head on this one... glad I finally found an answer... code is stupid! Giving up?!