Climbing svg + information text... but with bugs.
Added the climbing svg and tried to make it work but have a lot of work to do. Then I also added the information text but the text doesn't want to show up above the svg... probably because the "z-index" property doesn't apply to "position: static" objects... but I'll have to look into it another day!